horizon/requirements.txt
# ============================================================
# SYSTEMATIC PORTFOLIO ENGINE — DEPENDENCIES
# Install: pip install -r requirements.txt
# ============================================================

# Core data
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
duckdb>=0.9.0

# Portfolio optimization
cvxpy>=1.4.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Data sources
requests>=2.31.0
aiohttp>=3.9.0

# Broker
alpaca-trade-api>=3.0.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Dashboard
streamlit>=1.32.0
plotly>=5.18.0
